search: move alert observer out of graphqlbackend package
